b2b options for bnpl gateways
- bnpl gateways allow different shipping address
- bnpl „secured invoice“ allows b2b payments
