+++ src/Payone/Gateway/SecuredInvoice.php
@@ -2,8 +2,6 @@
 
 namespace Payone\Gateway;
 
-use Payone\Payone\Api\TransactionStatus;
-
 class SecuredInvoice extends PaylaBase {
 
 	const GATEWAY_ID = 'payone_secured_invoice';
@@ -17,6 +15,7 @@ class SecuredInvoice extends PaylaBase {
 		$this->icon               = PAYONE_PLUGIN_URL . 'assets/icon-rechnungskauf.png';
 		$this->method_title       = 'PAYONE ' . __( 'Secured Invoice', 'payone-woocommerce-3' );
 		$this->method_description = '';
+		$this->hide_when_b2b      = false;
 	}
 
 	public function init_form_fields() {
@@ -26,6 +25,8 @@ class SecuredInvoice extends PaylaBase {
 		];
 		$this->init_common_form_fields( 'PAYONE ' . __( 'Secured Invoice', 'payone-woocommerce-3' ) );
 		$this->form_fields['countries']['default'] = [ 'DE', 'AT' ];
+		$this->add_allow_different_shipping_address_field();
+		$this->form_fields['allow_different_shopping_address']['description'] = __( 'Attention: has to be enabled in the PAYONE account', 'payone-woocommerce-3' );
 	}
 
 	public function payment_fields() {
