+++ src/Payone/Transaction/Base.php
@@ -115,6 +115,24 @@ class Base extends Request {
 	}
 
 	/**
+	 * Sets the proper business relation of the customer according
+	 * to the provided billing address data.
+	 *
+	 * @param \WC_Order $order
+	 */
+	protected function set_business_relation( \WC_Order $order ) {
+		$company = $order->get_billing_company();
+
+		// Set b2b if billing company is present, set b2c otherwise
+		$this->set(
+			'businessrelation',
+			is_string( $company ) && ! empty( $company )
+				? 'b2b'
+				: 'b2c'
+		);
+	}
+
+	/**
 	 * "YYYY-MM-DD" -> "YYYYMMDD"
 	 *
 	 * @param string $birthday
