+++ src/Payone/Transaction/SafeInvoice.php
@@ -31,22 +31,4 @@ class SafeInvoice extends Base {
 
 		return $this->submit();
 	}
-
-	/**
-	 * Sets the proper business relation of the customer according
-	 * to the provided billing address data.
-	 *
-	 * @param \WC_Order $order
-	 */
-	protected function set_business_relation( \WC_Order $order ) {
-		$company = $order->get_billing_company();
-
-		// Set b2b if billing company is present, set b2c otherwise
-		$this->set(
-			'businessrelation',
-			is_string( $company ) && ! empty( $company )
-				? 'b2b'
-				: 'b2c'
-		);
-	}
 }
